refactor: adapt mongodb springdata new structure
