+++ mongodb/mongodb-springdata/src/main/java/io/flamingock/examples/mongodb/springboot/springdata/MongodbSpringbootSpringdata.java
@@ -19,6 +19,7 @@ package io.flamingock.examples.mongodb.springboot.springdata;
 import io.flamingock.api.annotations.EnableFlamingock;
 import io.flamingock.api.annotations.Stage;
 import io.flamingock.examples.mongodb.springboot.springdata.client.ClientRepository;
+import io.flamingock.examples.mongodb.springboot.springdata.config.MongoConfig;
 import io.flamingock.examples.mongodb.springboot.springdata.events.PipelineCompletedListener;
 import io.flamingock.examples.mongodb.springboot.springdata.events.PipelineFailedListener;
 import io.flamingock.examples.mongodb.springboot.springdata.events.PipelineStartedListener;
@@ -28,6 +29,7 @@ import io.flamingock.examples.mongodb.springboot.springdata.events.StageStartedL
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Import;
 import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
 
 
@@ -39,6 +41,7 @@ import org.springframework.data.mongodb.repository.config.EnableMongoRepositorie
 )
 @SpringBootApplication
 @EnableMongoRepositories(basePackageClasses = ClientRepository.class)
+@Import(MongoConfig.class)
 public class MongodbSpringbootSpringdata {
 
     public final static String DATABASE_NAME = "test";
