+++ mongodb/mongodb-springdata/src/test/java/io/flamingock/examples/mongodb/springboot/springdata/SpringData3SuccessExecutionTest.java
@@ -17,7 +17,6 @@
 package io.flamingock.examples.mongodb.springboot.springdata;
 
 
-import io.flamingock.community.mongodb.springdata.driver.SpringDataMongoAuditStore;
 import io.flamingock.examples.mongodb.springboot.springdata.config.MongoInitializer;
 import io.flamingock.examples.mongodb.springboot.springdata.events.PipelineCompletedListener;
 import io.flamingock.examples.mongodb.springboot.springdata.events.PipelineFailedListener;
@@ -25,7 +24,6 @@ import io.flamingock.examples.mongodb.springboot.springdata.events.PipelineStart
 import io.flamingock.examples.mongodb.springboot.springdata.events.StageCompletedListener;
 import io.flamingock.examples.mongodb.springboot.springdata.events.StageFailedListener;
 import io.flamingock.examples.mongodb.springboot.springdata.events.StageStartedListener;
-import io.flamingock.targetsystem.mongodb.springdata.MongoSpringDataTargetSystem;
 import org.bson.Document;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -71,11 +69,6 @@ class SpringData3SuccessExecutionTest {
     @Autowired
     private MongoTemplate mongoTemplate;
 
-    @Autowired
-    private MongoSpringDataTargetSystem mongoSpringDataTargetSystem;
-
-    @Autowired
-    private SpringDataMongoAuditStore auditStore;
 
     @Test
     @DisplayName("SHOULD create clientCollection and insert two clients")
